templated simple secondary vertex tag computer
